docs: add jsdoc (#189)
* docs: fix cli and compiler type references

* docs: fix missing core type references

* docs: fix missing plugins and extensions type references

* docs: fix more type references

* docs: fix cache type references

* docs: fix missing cloudflare type references

* docs: fix missing pubsub type references

* docs: fix util path typedoc warnings

* docs: add cli, core, generator-sdk jsdoc

* docs: change order of properties

* docs: add jsdoc to env and errors

* docs: add jsdoc to plugins

* docs: add jsdoc to typed pubsub

* docs: add auth extension to jsdoc

* docs: add complexity extension jsdoc

* fix: replace validation error with complexity error

* fix: remove @baeta/errors from complexity

* docs: add jsdocs for cache extensions

* docs: add jsdocs for cache extension stores

* fix build

* rename entry to index.md

* reorganize typedocs

* fix builds

* fix build

* chore: update docs

* chore: add jsdocs

+++ website/api-docs.mjs
@@ -1,28 +1,39 @@
 import fs from 'node:fs/promises';
 import * as TypeDoc from 'typedoc';
 
-async function cleanupDocs() {
-	await fs.rm('./docs/api/_media', { recursive: true, force: true });
-	await fs.unlink('./docs/api/README.md');
-}
+const noRm = process.argv.includes('--no-rm');
 
 async function generateDocs() {
-	await fs.rm('./docs/api', { recursive: true, force: true });
+	if (!noRm) {
+		await fs.rm('./docs/api', { recursive: true, force: true });
+	}
 
 	const packagesDir = await fs.readdir('../packages', { withFileTypes: true });
 	const dirs = packagesDir.filter((dirent) => dirent.isDirectory()).map((dirent) => dirent.name);
 
 	const app = await TypeDoc.Application.bootstrapWithPlugins({
+		out: './docs/api',
+		entryFileName: 'index.md',
 		entryPoints: dirs.map((dir) => `../packages/${dir}/`),
 		entryPointStrategy: 'packages',
 		plugin: ['typedoc-plugin-markdown'],
+		outputFileStrategy: 'modules',
+		readme: 'none',
 		hidePageHeader: true,
 		hideBreadcrumbs: true,
 		excludeExternals: true,
 		excludeScopesInPaths: true,
 		disableSources: true,
 		formatWithPrettier: true,
-		out: './docs/api',
+		parametersFormat: 'htmlTable',
+		propertiesFormat: 'htmlTable',
+		typeDeclarationFormat: 'htmlTable',
+		enumMembersFormat: 'htmlTable',
+		pageTitleTemplates: {
+			index: (args) => args.projectName,
+			member: (args) => args.name,
+			module: (args) => args.name,
+		},
 	});
 
 	const project = await app.convert();
@@ -32,10 +43,6 @@ async function generateDocs() {
 	}
 
 	await app.generateOutputs(project);
-
-	await cleanupDocs();
-
-	await fs.rename('./docs/api/packages.md', './docs/api/README.md');
 }
 
 generateDocs().catch(console.error);
